Integareted monocromatic colors + changed the layout of graphs for customers into graph-line charts
