+++ src/admin/components/KPI.tsx
@@ -1,5 +1,4 @@
 import { Button, clx, Container, Text } from '@medusajs/ui';
-import { useCustomerAnalytics } from '../hooks/customer-analytics';
 import { SmallCardSkeleton } from '../skeletons/SmallCardSkeleton';
 import { LineChart } from './LineChart';
 import { Skeleton } from './Skeleton';
@@ -240,114 +239,6 @@ export const TotalOrders = () => {
                 xAxisDataKey="name"
                 yAxisDataKey="count"
                 lineColor="#a1a1aa"
-                hideTooltip
-              />
-            </div>
-          </div>
-        </div>
-      )}
-    </Container>
-  );
-};
-
-export const AverageSalesPerCustomer = () => {
-  const { interval, range } = useIntervalRange();
-  const ordersQuery = useOrderAnalytics(interval, range);
-  const customersQuery = useCustomerAnalytics(range);
-
-  const currencyCode = (
-    ordersQuery.data?.currency_code ||
-    customersQuery.data?.currency_code ||
-    'EUR'
-  ).toUpperCase();
-
-  const averageSalesPerCustomer =
-    customersQuery.data?.total_customers &&
-    customersQuery.data.total_customers > 0
-      ? +(
-          (ordersQuery.data?.total_sales || 0) /
-          customersQuery.data.total_customers
-        ).toFixed(2)
-      : 0;
-
-  const customersPerBucket = new Map(
-    (customersQuery.data?.customer_count ?? []).map((point) => [
-      point.name,
-      (point.new_customers ?? 0) + (point.returning_customers ?? 0),
-    ]),
-  );
-
-  const averageSalesPerCustomerTimeline = (
-    ordersQuery.data?.order_sales ?? []
-  ).map((point) => {
-    const customersInBucket =
-      customersPerBucket.get(point.name) ??
-      customersQuery.data?.total_customers ??
-      0;
-
-    return {
-      name: point.name,
-      value:
-        customersInBucket > 0
-          ? Number((point.sales / customersInBucket).toFixed(2))
-          : 0,
-    };
-  });
-  const isLoading = ordersQuery.isLoading || customersQuery.isLoading;
-
-  return (
-    <Container className="flex flex-col">
-      <div className="flex justify-between items-center">
-        <Text size="large">Average Sales per Customer</Text>
-        <a
-          href={withOptionalAnalyticsRange(
-            '/app/analytics#:~:text=Sales%20Over%20Time',
-            range,
-          )}
-        >
-          <Button variant="transparent" className="text-ui-fg-muted text-xs">
-            View more
-          </Button>
-        </a>
-      </div>
-      {isLoading ? (
-        <div className="flex gap-4 justify-between flex-1">
-          <div>
-            <SmallCardSkeleton />
-          </div>
-          <Skeleton className="aspect-video w-64 mt-2.5" />
-        </div>
-      ) : (
-        <div className="flex gap-4 justify-between flex-1">
-          <div>
-            <Text size="xlarge" weight="plus">
-              {new Intl.NumberFormat(undefined, {
-                currency: currencyCode,
-                style: 'currency',
-              }).format(averageSalesPerCustomer)}
-            </Text>
-            <KPITimelineLabel percentage={averageSalesPerCustomer / 100} />
-          </div>
-          <div className="flex-1 flex mt-2.5">
-            <div className="aspect-video mt-auto w-full max-w-64 ml-auto">
-              <LineChart
-                data={averageSalesPerCustomerTimeline}
-                xAxisDataKey="name"
-                yAxisDataKey="value"
-                lineColor="#a1a1aa"
-                yAxisTickFormatter={(value) =>
-                  new Intl.NumberFormat(undefined, {
-                    style: 'currency',
-                    currency: currencyCode,
-                    maximumFractionDigits: 0,
-                  }).format(
-                    typeof value === 'number'
-                      ? value
-                      : typeof value === 'string'
-                        ? Number(value)
-                        : 0,
-                  )
-                }
                 hideTooltip
               />
             </div>
